pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "droidbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "filelock>=3.20.0",
    "networkx>=3.6",
    "adbutils>=2.9.1,<3.0.0",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "setuptools>=80.9.0",
    "uiautomator2==3.3.3",
    "numpy>=1.23.0",
]

[project.optional-dependencies]
web = [
    "pyscrcpy>=1.0.1",
    "av>=9.0.0",
    "natsort>=8.4.0",
    # MySQL 用户管理功能依赖
    "aiofiles==23.2.1",
    "fastapi>=0.110.0",
    "fastapi-sessions>=0.3.2",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.24.0",
    "sqlmodel>=0.0.14",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "volcengine-python-sdk>=5.0.6",
    "onnxruntime>=1.14.0",
]

yolomode = [
    "onnxruntime>=1.14.0",
]

[tool.setuptools]
packages = ["droidbot", "droidbot.adapter"]

[tool.uv]
override-dependencies = [
    "adbutils>=2.9.1,<3.0.0",
    "av>=9.0.0"
]
allow-insecure-host = ["pypi.hobot.cc", "mirrors.aliyun.com"]

[[tool.uv.index]]
url = "http://pypi.hobot.cc/hobot-local/simple"
default = true

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple/"
